# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MontoCRM API
  version: 2.0.0
  description: |
    AI 原生 B2B 大客户 CRM — Phase 2（P5–P9）契约。

    实现对照：
    - P5 Guard hybrid + `/admin/llm/guard/test`；拦截审计 `guard.block`
    - P6 Langfuse Mock + `/metrics` + Ragas eval
    - P7 混合检索 + `/admin/rag/test-retrieval` / reindex
    - P8 钉钉自定义机器人 + `/admin/notify/dingtalk`
    - P9 CI、Nginx TLS/限流、安全头；`/auth/token` 与 `/activities/extract` 可返回 429

    运行时 FastAPI `version` 为 2.0.0。登录实现为 OAuth2 password（`/auth/token`，`/auth/login` 为别名）。
  contact:
    name: MontoCRM
servers:
  - url: http://localhost:8000
    description: Local API（无 /api 前缀；经 Nginx 时为 https://host/api/）
  - url: https://localhost/api
    description: Compose Nginx TLS
...
paths:
...
  /activities:
    get:
      tags: [Activities]
      summary: 活动列表
      operationId: listActivities
      security:
        - bearerAuth: []
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: opportunity_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
    post:
      tags: [Activities]
      summary: 创建活动（手工）
      operationId: createActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityCreate'
      responses:
        '201':
          description: Created
  /activities/extract:
    post:
      tags: [Activities, Writeback]
      summary: 提交会议纪要并触发 Orchestrator 并行分析（闭环 A）
      operationId: extractFromNotes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractRequest'
      responses:
        '200':
          description: 已生成写回建议，等待确认
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritebackProposal'
        '429':
          description: 抽取过频（生产 10 req/min/用户；Nginx 另有 limit_req）
  /writeback/{proposal_id}/confirm:
    post:
      tags: [Writeback]
      summary: 确认写回（HITL L1）
      operationId: confirmWriteback
      security:
        - bearerAuth: []
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WritebackConfirmRequest'
      responses:
        '200':
          description: 写回成功
  /writeback/{proposal_id}/reject:
    post:
      tags: [Writeback]
      summary: 拒绝写回
      operationId: rejectWriteback
      security:
        - bearerAuth: []
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 已保存原始 Activity
...
  /pending-actions/{pending_id}:
    get:
      tags: [PendingActions]
      summary: 获取待确认动作
      operationId: getPendingAction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PendingId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingAction'
  /pending-actions/{pending_id}/confirm:
    post:
      tags: [PendingActions]
      summary: 确认待处理动作（HITL L1/L2）
      operationId: confirmPendingAction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PendingId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PendingConfirmRequest'
      responses:
        '200':
          description: 已执行
  /pending-actions/{pending_id}/reject:
    post:
      tags: [PendingActions]
      summary: 拒绝待处理动作
      operationId: rejectPendingAction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PendingId'
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PendingId:
      name: pending_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ActivityCreate:
      type: object
      required: [account_id, type, content]
      properties:
        account_id:
          type: string
          format: uuid
        opportunity_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [meeting, call, email, note]
        subject:
          type: string
        content:
          type: string
        occurred_at:
          type: string
          format: date-time
    ExtractRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 50
        account_id:
          type: string
          format: uuid
        opportunity_id:
          type: string
          format: uuid
    WritebackProposal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opportunity_updates:
          type: object
        contact_updates:
          type: array
          items:
            type: object
        new_contacts:
          type: array
          items:
            type: object
        tasks:
          type: array
          items:
            type: object
        reasoning:
          type: string
        evidence:
          type: array
          items:
            type: object
    WritebackConfirmRequest:
      type: object
      properties:
        accepted_fields:
          type: object
          description: 用户勾选/修改后的最终字段
...
    PendingAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [writeback, send_email, stage_change, amount_change]
        hitl_level:
          type: string
          enum: [L1, L2]
        status:
          type: string
          enum: [pending, confirmed, rejected, expired]
        payload:
          type: object
        expires_at:
          type: string
          format: date-time
    PendingConfirmRequest:
      type: object
      properties:
        edited_payload:
          type: object
...